requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
pyjwt==2.8.0
psutil==5.9.8
